Fix a compiler warning.
git-svn-id: svn://svn.berlios.de/gale/trunk@1317 2d798ef8-6e06-0410-941a-9092aa770104

+++ gale2/demo/heartbeat/bg_frag.glsl
@@ -1,4 +1,4 @@
-uniform vec2 viewport;
+uniform ivec2 viewport;
 uniform int sections;
 uniform vec4 color0;
 uniform vec4 color1;
